Move more components into correct folders
