+++ src/components/Filter/PunchListFilter/usePunchListFilterFacade.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useEffect } from 'react';
 import { useState } from 'react';
-import { PunchPreview } from '../../../services/apiTypes';
+import { PunchPreview } from '../../../typings/apiTypes';
 import useFilterFacade, { Filter } from '../useFilter';
 import { filterPunchPreviews } from './filterPunchPreviews';
