+++ src/index.tsx
@@ -16,10 +16,10 @@ import SkeletonLoadingPage from './components/loading/SkeletonLoader';
 import removeSubdirectories from './utils/removeSubdirectories';
 import EntityDetails from './components/EntityDetails/EntityDetails';
 import TextIcon from './components/EntityDetails/TextIcon';
-import PunchList from './PunchList/PunchList';
-import Scope from './Scope/Scope';
-import FooterButton from './Footer/FooterButton';
-import NavigationFooter from './Footer/NavigationFooter';
+import PunchList from './components/PunchList';
+import Scope from './components/Scope';
+import FooterButton from './components/Footer/FooterButton';
+import NavigationFooter from './components/Footer/NavigationFooter';
 import SearchTypeButton from './components/buttons/SearchTypeButton';
 import { StorageKey } from './typings/enums';
 import PageHeader from './components/PageHeader';
